rewrite logging, change some defaults, remove airflow args and functions, add requirements

+++ rcsb/workflow/bcif/task_functions.py
@@ -13,7 +13,7 @@ from rcsb.workflow.bcif.bcif_functions import bcifconvert
 from mmcif.api.DictionaryApi import DictionaryApi
 from mmcif.io.IoAdapterPy import IoAdapterPy as IoAdapter
 
-logging.basicConfig(level=logging.INFO)
+logger = logging.getLogger(__name__)
 
 def status_start_(list_file_base: str, status_start_file: str) -> bool:
     start_file = os.path.join(list_file_base, status_start_file)
@@ -46,7 +46,7 @@ def make_dirs_(workflow_utility: WorkflowUtilities) -> bool:
             os.mkdir(path, mode=0o777)
     return True
 
-def get_pdb_list_(workflow_utility: WorkflowUtilities, load_type: str, list_file_base: str, pdb_list_filename: str, result: bool) -> bool:
+def get_pdb_list_(workflow_utility: WorkflowUtilities, load_type: str, list_file_base: str, pdb_list_filename: str, result=None) -> bool:
     outfile = os.path.join(list_file_base, pdb_list_filename)
     if os.path.exists(outfile):
         return True
@@ -57,7 +57,7 @@ def get_pdb_list_(workflow_utility: WorkflowUtilities, load_type: str, list_file
         pickle.dump(pdb_list, w)
     return True
 
-def get_csm_list_(workflow_utility: WorkflowUtilities, load_type: str, list_file_base: str, csm_list_filename: str, result: bool) -> bool:
+def get_csm_list_(workflow_utility: WorkflowUtilities, load_type: str, list_file_base: str, csm_list_filename: str, result=None) -> bool:
     outfile = os.path.join(list_file_base, csm_list_filename)
     if os.path.exists(outfile):
         return True
@@ -71,13 +71,13 @@ def get_csm_list_(workflow_utility: WorkflowUtilities, load_type: str, list_file
     return True
 
 def make_task_list_from_remote_(list_file_base: str, pdb_list_filename: str, csm_list_filename: str,
-                                input_list_filename: str, maxfiles: int, workflow_utility: WorkflowUtilities, result: bool) -> bool:
+                                input_list_filename: str, maxfiles: int, workflow_utility: WorkflowUtilities, result=None) -> bool:
     # read pdb list
     pdb_list = None
     with open(os.path.join(list_file_base, pdb_list_filename), "rb") as r:
         pdb_list = pickle.load(r)
     if not pdb_list:
-        logging.exception("error reading pdb list")
+        logger.error("error reading pdb list")
         return False
     # read csm list
     csm_list = None
@@ -86,26 +86,26 @@ def make_task_list_from_remote_(list_file_base: str, pdb_list_filename: str, csm
         with open(csm_list_path, "rb") as r:
             csm_list = pickle.load(r)
     if not csm_list:
-        logging.exception("error reading csm list")
+        logger.error("error reading csm list")
     else:
         # join lists
         pdb_list.extend(csm_list)
     # trim list if testing
     nfiles = len(pdb_list)
-    logging.info("found %d cif files" % nfiles)
+    logger.info("found %d cif files" % nfiles)
     if nfiles == 0:
         return False
     if maxfiles > 0 and maxfiles < nfiles:
         nfiles = maxfiles
         pdb_list = pdb_list[0:nfiles]
-        logging.info("reading only %d files" % nfiles)
+        logger.info("reading only %d files" % nfiles)
     # save input list
     outfile = os.path.join(list_file_base, input_list_filename)
     with open(outfile, "wb") as w:
         pickle.dump(pdb_list, w)
     return True
 
-def make_task_list_from_local_(local_data_path: str, list_file_base: str, input_list_filename: str, result: bool) -> bool:
+def make_task_list_from_local_(local_data_path: str, list_file_base: str, input_list_filename: str, result=None) -> bool:
     """
     requires cif files in source folder with no subdirs
     writes to target folder with no subdirs
@@ -116,44 +116,44 @@ def make_task_list_from_local_(local_data_path: str, list_file_base: str, input_
     if nfiles == 0:
         tasklist = glob.glob(os.path.join(local_data_path, "*.cif"))
         nfiles = len(tasklist)
-    logging.info("found %d cif files" % nfiles)
+    logger.info("found %d cif files" % nfiles)
     with open(os.path.join(list_file_base, input_list_filename), "wb") as w:
         pickle.dump(tasklist, w)
     return True
 
-def split_tasks_(list_file_base: str, input_list_filename: str, input_list_2d: str, maxfiles: int, subtasks:int, result: bool) -> List[int]:
+def split_tasks_(list_file_base: str, input_list_filename: str, input_list_2d: str, maxfiles: int, subtasks:int, result=None) -> List[int]:
     # read task list
     tasklist = None
     with open(os.path.join(list_file_base, input_list_filename), "rb") as r:
         tasklist = pickle.load(r)
     if not tasklist:
-        logging.exception("error reading task list")
+        logger.error("error reading task list")
         return None
     # trim list
     nfiles = len(tasklist)
-    logging.info("found %d cif files" % nfiles)
+    logger.info("found %d cif files" % nfiles)
     if nfiles == 0:
         return []
     if maxfiles > 0 and maxfiles < nfiles:
         nfiles = maxfiles
-        logging.info("reading only %d files" % nfiles)
+        logger.info("reading only %d files" % nfiles)
     # divide into subtasks
     if (subtasks is None) or not str(subtasks).isdigit():
         subtasks = 1
     subtasks = int(subtasks)
     if subtasks == 0:
         subtasks = multiprocessing.cpu_count()
-        logging.info("machine has %d processors" % subtasks)
+        logger.info("machine has %d processors" % subtasks)
     else:
-        logging.info("dividing across %d subtasks" % subtasks)
+        logger.info("dividing across %d subtasks" % subtasks)
     tasks = split_list(nfiles, subtasks, tasklist)
     # save full tasks file
-    logging.info("get local tasks saving %d tasks to %s" % (len(tasks), input_list_2d))
+    logger.info("get local tasks saving %d tasks to %s" % (len(tasks), input_list_2d))
     with open(os.path.join(list_file_base, input_list_2d), "wb") as w:
         pickle.dump(tasks, w)
     # return list of task indices
     tasks = list(range(0, len(tasks)))
-    logging.info("returning %d tasks" % len(tasks))
+    logger.info("returning %d tasks" % len(tasks))
     return tasks
 
 def split_list(nfiles:int, subtasks:int, tasklist:List[str]) -> List[List[str]]:
@@ -161,7 +161,7 @@ def split_list(nfiles:int, subtasks:int, tasklist:List[str]) -> List[List[str]]:
     if step < 1:
         step = 1
     steps = nfiles // step
-    logging.info("split list has %d files and %d steps with step %d" % (nfiles, steps, step))
+    logger.info("split list has %d files and %d steps with step %d" % (nfiles, steps, step))
     if not isinstance(tasklist[0], str):
         tasklist = [str(task) for task in tasklist]
     tasks = [tasklist[index*step:step + index*step] if index < steps - 1 else tasklist[index*step:nfiles] for
@@ -178,9 +178,9 @@ def local_task_map_(index: int, list_file_base: str, input_list_2d: str, temp_pa
         allfiles = pickle.load(r)
         infiles = allfiles[index]
     if not infiles:
-        logging.exception("error - no infiles")
+        logger.error("error - no infiles")
         return False
-    logging.info("task map has %d infiles" % len(infiles))
+    logger.info("task map has %d infiles" % len(infiles))
 
     # form dictionary object
     da = None
@@ -193,7 +193,7 @@ def local_task_map_(index: int, list_file_base: str, input_list_2d: str, temp_pa
                 containers += adapter.readFile(inputFilePath=path)
             da = DictionaryApi(containerList=containers, consolidate=True)
         except Exception as e:
-            logging.exception("error - failed to create dictionary api")
+            logger.exception("error - failed to create dictionary api")
 
     def batch_task(tasks, local_inputs_or_remote, update_base, compress, temp_path, workflow_utility, python_molstar_java, da, molstar_cmd):
         for task in tasks:
@@ -211,14 +211,14 @@ def local_task_map_(index: int, list_file_base: str, input_list_2d: str, temp_pa
                 if not os.path.exists(cif_file_path):
                     cif_file_path = "%s.gz" % cif_file_path
                     if not os.path.exists(cif_file_path):
-                        logging.exception("error - could not find %s" % cif_file_path)
+                        logger.error("error - could not find %s" % cif_file_path)
                         return
             pdb_file_name = os.path.basename(cif_file_path)
             bcif_file_path = os.path.join(update_base,
                                             pdb_file_name.replace(".cif.gz", ".bcif").replace(".cif", ".bcif"))
             if compress:
                 bcif_file_path = "%s.gz" % bcif_file_path
-            logging.info("converting %s to %s" % (cif_file_path, bcif_file_path))
+            logger.info("converting %s to %s" % (cif_file_path, bcif_file_path))
         else:
             tokens = line.split(' ')
             pdb_id = tokens[0]
@@ -247,12 +247,12 @@ def local_task_map_(index: int, list_file_base: str, input_list_2d: str, temp_pa
                 else:
                     raise requests.exceptions.RequestException("error - request failed for %s" % url)
             except Exception as e:
-                logging.exception(str(e))
+                logger.exception(str(e))
                 if os.path.exists(cif_file_path):
                     os.unlink(cif_file_path)
                 return
         if os.path.exists(bcif_file_path):
-            logging.info("file %s already exists" % bcif_file_path)
+            logger.info("file %s already exists" % bcif_file_path)
             if local_inputs_or_remote == "remote":
                 os.unlink(cif_file_path)
             return
@@ -270,7 +270,7 @@ def local_task_map_(index: int, list_file_base: str, input_list_2d: str, temp_pa
             shutil.chown(bcif_file_path, 'root', 'root')
             os.chmod(bcif_file_path, 0o777)
         except Exception as e:
-            logging.exception(str(e))
+            logger.exception(str(e))
         # remove input file
         finally:
             if local_inputs_or_remote == "remote":
@@ -305,7 +305,7 @@ def local_task_map_(index: int, list_file_base: str, input_list_2d: str, temp_pa
     return True
 
 
-def validate_output_(list_file_base:str, input_list_filename:str, update_base:str, compress:bool, missing_file_base:str, missing_filename:str, wu:WorkflowUtilities) -> bool:
+def validate_output_(list_file_base:str, input_list_filename:str, update_base:str, compress:bool, missing_file_base:str, missing_filename:str, wu:WorkflowUtilities, result=None) -> bool:
     input_list_file = os.path.join(list_file_base, input_list_filename)
     if not os.path.exists(input_list_file):
         return False
@@ -331,25 +331,30 @@ def validate_output_(list_file_base:str, input_list_filename:str, update_base:st
     return True
 
 
-def remove_temp_files_(temp_path:str) -> bool:
+def remove_temp_files_(temp_path:str, list_file_base:str, result=None) -> bool:
     if not os.path.exists(temp_path):
         return False
     try:
         for filename in os.listdir(temp_path):
             path = os.path.join(temp_path, filename)
-            os.unlink(path)
+            if os.path.isfile(path):
+                os.unlink(path)
+        for filename in os.listdir(list_file_base):
+            path = os.path.join(list_file_base, filename)
+            if os.path.isfile(path):
+                os.unlink(path)
     except Exception as e:
-        logging.warning(str(e))
+        logger.warning(str(e))
     return True
 
 
-def tasks_done_(local_task_maps: list) -> bool:
-    logging.info("task maps completed")
+def tasks_done_(result=None) -> bool:
+    logger.info("task maps completed")
     return True
 
 
 def k8s_branch_() -> bool:
-    logging.info("using k8s tasks")
+    logger.info("using k8s tasks")
     return True
